+++ src/steps/intune/index.ts
@@ -1,3 +1,7 @@
+import { deviceConfigurationAndFindingsSteps } from './steps/deviceConfigurationsAndFindings';
 import { deviceSteps } from './steps/devices';
 
-export const intuneSteps = [...deviceSteps];
+export const intuneSteps = [
+  ...deviceSteps,
+  ...deviceConfigurationAndFindingsSteps,
+];