+++ src/steps/intune/steps/devices/converters.ts
@@ -1,5 +1,4 @@
 import {
-  createDirectRelationship,
   createIntegrationEntity,
   createMappedRelationship,
   Entity,
@@ -8,12 +7,12 @@ import {
 } from '@jupiterone/integration-sdk-core';
 import { ManagedDevice } from '@microsoft/microsoft-graph-types-beta';
 import { entities as activeDirectoryEntities } from '../../../active-directory';
-import { entities, relationships } from './constants';
-
-type ManagedDeviceEntity = Entity;
+import { entities, relationships } from '../../constants';
 
 // https://docs.microsoft.com/en-us/graph/api/resources/intune-devices-manageddevice?view=graph-rest-1.0&viewFallbackFrom=graph-rest-beta
-export function createManagedDeviceEntity(managedDevice: ManagedDevice) {
+export function createManagedDeviceEntity(
+  managedDevice: ManagedDevice,
+): Entity {
   return createIntegrationEntity({
     entityData: {
       source: managedDevice,
@@ -86,7 +85,7 @@ export function createManagedDeviceEntity(managedDevice: ManagedDevice) {
  * Should be used when the active directory is not being ingested in this integration instance.
  */
 export function createUserDeviceMappedRelationship(
-  managedDeviceEntity: ManagedDeviceEntity,
+  managedDeviceEntity: Entity,
   userId?: string,
   email?: string,
 ): Relationship | undefined {
@@ -111,15 +110,4 @@ export function createUserDeviceMappedRelationship(
         },
       })
     : undefined;
-}
-
-export function createUserDeviceDirectRelationship(
-  managedDeviceEntity: ManagedDeviceEntity,
-  userEntity: Entity,
-): Relationship {
-  return createDirectRelationship({
-    _class: relationships.USER_HAS_DEVICE._class,
-    from: userEntity,
-    to: managedDeviceEntity,
-  });
 }