+++ src/steps/intune/steps/devices/index.ts
@@ -1,14 +1,14 @@
 import {
+  createDirectRelationship,
   IntegrationStepExecutionContext,
   Step,
 } from '@jupiterone/integration-sdk-core';
 import { IntegrationConfig, IntegrationStepContext } from '../../../../types';
 import { steps as activeDirectorySteps } from '../../../active-directory';
 import { DeviceManagementIntuneClient } from '../../clients/deviceManagementIntuneClient';
-import { relationships, entities, steps } from './constants';
+import { relationships, entities, steps } from '../../constants';
 import {
   createManagedDeviceEntity,
-  createUserDeviceDirectRelationship,
   createUserDeviceMappedRelationship,
 } from './converters';
 
@@ -26,7 +26,11 @@ export async function fetchDevices(
     await jobState.addEntity(deviceEntity);
     const userEntity = await jobState.findEntity(device.userId as string);
     const userDeviceRelationship = userEntity
-      ? createUserDeviceDirectRelationship(deviceEntity, userEntity)
+      ? createDirectRelationship({
+          _class: relationships.USER_HAS_DEVICE._class,
+          from: userEntity,
+          to: deviceEntity,
+        })
       : createUserDeviceMappedRelationship(
           deviceEntity,
           device.userId as string,