+++ CHANGELOG.md
@@ -8,16 +8,25 @@ and this project adheres to
 
 ## [Unreleased]
 
-## 2.0. - 2020-02-10
+## 2.0.0 - 2020-03-01
 
 ### Added
 
-- Ingetsion of `Managed Device` entities
-- Ingestion of `UserHasDevice` relationships
+- Ingetsion of `Managed Device`, `Device Configuration`, and
+  `Noncompliance Finding` entities
+- Ingestion of `UserHasDevice`, `DeviceUsesDeviceConfiguration`,
+  `DeviceConfigurationIdentifiesNoncomplianceFinding` and
+  `DeviceHasNoncompliancceFinding` relationships
 
 ### Updated
 
 - Updated jupiterone.md documentation.
+
+### Changed
+
+- Changed Azure Active Directory entities and relationships to have a type of
+  `azure_` to match
+  [the Azure graph project](https://github.com/JupiterOne/graph-azure)
 
 ### Fixed